requirements.txt
face-recognition-models==0.3.*
Flask==1.0.*
gTTS==2.0.*
gTTS-token==1.1.*
numpy==1.16.*
Pillow==6.0.*
PyAudio==0.2.*
pygame==1.9.*
pyparsing==2.4.*
python-dateutil==2.8.*
pytz==2019.*
requests==2.22.*
requests-oauthlib==1.0.*
responses==0.9.*
soupsieve==1.9.*
SpeechRecognition==3.8.*
timeloop==1.0.*
xlwt==1.3.*
